Added navbar, refactored components, and updated CSS
