+++ src/app.css
@@ -1,0 +1,12 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+html {
+  background-color: rgb(240, 240, 240);
+}
+
+main {
+  margin: 2rem;
+  margin-left: 15rem;
+}
