+++ src/routes/admin/+page.svelte
@@ -1,6 +1,6 @@
 <main class='flex flex-col'>
   <div class='flex ml-4 mb-8 w-full bg-white rounded p-4'>
-    <h1 class='text-5xl text-violet-800'>Computer Science Capstone Presentation Tools</h1>
+    <h1 class='text-5xl text-violet-800'>Admin Tools</h1>
   </div>
   <div class='flex ml-4 mb-8 w-full bg-white rounded p-4'>
     Put some description here.
